refactor: move globe specific code to ApiGlobe
